# swagger/openapi.yaml
openapi: 3.0.3
info:
  title: Hard Allocation Platform API
  description: |
    Pharmaceutical Production Sequencing & Allocation Optimizer.
    v1 Core allocation · v2 Enterprise (rules, exceptions, risk, RBAC) ·
    v3 AI-Assisted Planning (copilot, agents, digital twin, predictive risk).
    AI recommends — human planner retains final approval.
  version: 3.0.0
  contact:
    name: Enterprise Architecture
servers:
  - url: http://localhost:8000/api/v1
    description: Core allocation API (MVP 1.0)
  - url: http://localhost:8000/api/v2
    description: Enterprise API (MVP 2.0) — requires X-User-Id, X-User-Role headers
  - url: http://localhost:8000/api/v3
    description: AI-Assisted Planning (MVP 3.0) — requires X-User-Id, X-User-Role headers
...
paths:
...
  /auth/login:
    post:
      tags: [Enterprise Auth]
      summary: Login with demo username
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                username: { type: string, example: planner }
      responses:
        '200':
          description: User profile with permissions
  /auth/me:
    get:
      tags: [Enterprise Auth]
      summary: Current user profile
      parameters:
        - in: header
          name: X-User-Id
          schema: { type: string }
        - in: header
          name: X-User-Role
          schema: { type: string, enum: [PLANNER, QA, SUPPLY_CHAIN, ADMIN, VIEWER] }
      responses:
        '200':
          description: User with role permissions
  /rules:
    get:
      tags: [Enterprise Rules]
      summary: List versioned rules (v2)
      parameters:
        - in: query
          name: category
          schema: { type: string, enum: [COUNTRY, CUSTOMER, PRODUCT, MARKET, SEQUENCE, RMSL, BATCH_SPLIT] }
      responses:
        '200':
          description: Rules with categories and audit log
    post:
      tags: [Enterprise Rules]
      summary: Create rule (ADMIN)
      responses:
        '201':
          description: Created rule
...
  /rules/{ruleId}:
    get:
      tags: [Enterprise Rules]
      summary: Get rule with version history
      parameters:
        - in: path
          name: ruleId
          required: true
          schema: { type: string }
      responses:
        '200':
          description: Rule detail
    put:
      tags: [Enterprise Rules]
      summary: Update rule (creates new version)
      parameters:
        - in: path
          name: ruleId
          required: true
          schema: { type: string }
      responses:
        '200':
          description: Updated rule
  /exceptions:
    get:
      tags: [Exceptions]
      summary: Exception queue
      parameters:
        - in: query
          name: status
          schema: { type: string }
        - in: query
          name: type
          schema: { type: string }
      responses:
        '200':
          description: Exception list
  /exceptions/{exceptionId}/comments:
    post:
      tags: [Exceptions]
      summary: Add comment to exception
      parameters:
        - in: path
          name: exceptionId
          required: true
          schema: { type: string }
      responses:
        '200':
          description: Updated exception
  /exceptions/{exceptionId}/review:
    post:
      tags: [Exceptions]
      summary: Mark exception in review
      parameters:
        - in: path
          name: exceptionId
          required: true
          schema: { type: string }
      responses:
        '200':
          description: Updated exception
  /exceptions/{exceptionId}/escalate:
    post:
      tags: [Exceptions]
      summary: Escalate exception
      parameters:
        - in: path
          name: exceptionId
          required: true
          schema: { type: string }
      responses:
        '200':
          description: Escalated exception
  /exceptions/{exceptionId}/resolve:
    post:
      tags: [Exceptions]
      summary: Resolve exception
      parameters:
        - in: path
          name: exceptionId
          required: true
          schema: { type: string }
      responses:
        '200':
          description: Resolved exception
  /jobs/mass-allocation:
    post:
      tags: [Mass Jobs]
      summary: Start background mass allocation job
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                period: { type: string, enum: [DAILY, WEEKLY], default: DAILY }
                execute: { type: boolean, default: false }
                orderIds: { type: array, items: { type: string } }
      responses:
        '202':
          description: Job queued
...
  /jobs/{jobId}:
    get:
      tags: [Mass Jobs]
      summary: Get job with progress and results
      parameters:
        - in: path
          name: jobId
          required: true
          schema: { type: string }
      responses:
        '200':
          description: Job detail with result history
...
  /copilot/ask:
    post:
      tags: [AI Planning]
      summary: Planning Copilot v3 — explainable business answers
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                question: { type: string }
                packagingOrderId: { type: string, example: PO-20001 }
      responses:
        '200':
          description: Answer with evidence and graph context
...
components:
  securitySchemes:
    RoleHeaders:
      type: apiKey
      in: header
      name: X-User-Role
...